Alexa Commands NLP
Project Completed at the University of Colorado at Boulder
Assignment Description:
Let's pretend your a business analyst at Amazon. Amazon is aggressively trying to monetize advertising space on its growing army of devices. You're focused on creating advertising opportunities on one of Amazon's most popular products, the Echo Show.

The Echo Show is designed to display articles that are related to the queries that consumers make with it. For instance, if people ask Alexa to "play NPR news," Alexa shows articles related to politics. Amazon wants to grow this article suggestion feature to advertising. That is, they want to give advertisers the ability to advertise with contextual relevance. If a consumer asks Amazon for Recipes, for instance, Whole Foods might want to advertise a recipe chocked full of recipe items available at local whole foods with a "shop at Whole Foods" button.

Your goal is to help Amazon to brainstorm what different types of advertising content consumers might find closely related to their queries. To do so, you've asked an engineer to give you a sample of Echo Show queries, that is, a set of things that people have asked Alexa.

The Data:
You've been given a plain-text file of 10,581 different Amazon Echo Show queries. Download it here Each line is a query. Basic file processing techniques in Python should make this data very easy to work with.