GEO Locating Twitter Data
This is a straightforward assignment in which you will:

Extract self-declared users' locations from their Tweets
Use the external package geostring to resolve the locations to known cities, states, and countries.

We are interested in the geographic distribution of Tweets around a particular brand. In this case, we are looking to find where people are located who are Tweeting about Mac N Cheese Breakfast.

Scrutinizing the Tweet data structure
At this point, you would probably consider using the Twitter API to fetch the data you need. We have done the actual data fetching for you, but you will need to understand the Twitter API data structure in order to do this assignment.

Take a look at all the metadata associated with a Tweet. It's a lot: https://developer.twitter.com/en/docs/tweets/data-dictionary/overview/tweet-object (Links to an external site.)

Find the field that maps to user reported location, that is, this field:


This is the field you will be extracting from each Tweet. Because this field is a self-reported location, the data is a bit fuzzy. We will run this location through the geostring library to normalize it to known locations.

Take a moment to:
Review the Tweet data structure and to find the field representing the user's location
Familiarize yourself with the geostring library.

The Data
To get this kind of Data, you would normally interface with the Twitter API. Here, we have done that work for you, hitting the Twitter API's search endpoint and fetching 10 pages for a total of 1000 Tweets.