Extracting Keywords from Sponsored Content
You will fetch archived sponsored content from archive.org and will extract the keywords from the content.

Sponsored content on news publishers sites are native in the sense that the look like a news story. They're a popular new type of ad for this reason. Let's say we're a top marketing agency and we wanted to get a better understanding of what types of sponsored content articles exist on major publisher sites. We searched Google for urls that appeared to indicate in the URL that the story was sponsored (e.g., www.nytimes.com/sponsored-story/story-url.html). However sponsored content is elusive in that it tends to disappear. Whenever we want to see what was on a webpage in the past, we can search archive.org to find web pages that no longer exist. For this challenge, go through the list of dead URLs and use archive.org to retrieve as many of the news stories as you can.

The tasks in this assignment are straightforward, but you will need to understand how to use a few 3rd party libraries to get this done. Glance at the documentation for the tools listed below, and come back to them as you work through the assignment.

Requests: HTTP for Humans
Requests is a library for fetching http content. Alternatively, you could use Python's urllib, althought most people find requests to be easier to work with.

Newspaper: Article scraping & curation
Newspaper is a full framework-like utility for managing the fetching of online news-oriented content. We will not be using all the rich features it contains. Rather, we are interested in Newspaper for its straightforward ability to (generally) extract the primary text from a web page. To this end, you will use requests to fetch the pages, and will pass the html content to Newspaper's fulltext method to extract the text.

Important: When you install, please be sure to install newspaper3k, not newspaper.

RAKE: A keyword extraction algorithm
RAKE is a general algorithm for keyword extraction. Here, we will be using a specific implementation that is packaged into the library called python-rake.

You will use rake to extract keywords (key-phrases actually) from the contents of the pages after reducing the HTML to plain extracted text content.